requirements.txt
PyQt6>=6.6.0
cryptography>=41.0.0
argon2-cffi>=23.1.0
pyotp>=2.9.0
pytest>=7.4.0
pytest-cov>=4.1.0
